+++ src/components/Popper/index.js
@@ -1,7 +1,0 @@
-export {
-  Reference as PopperReference,
-  Manager as PopperManager,
-  placements as PopperPlacements,
-} from 'react-popper'
-export { default as Popper } from './Popper'
-export { default as minWidthModifier } from './minWidthModifier'