+++ src/components/index.js
@@ -1,5 +1,3 @@
 export { default as ClickOutside } from './ClickOutside'
-export { default as Dropdown } from './Dropdown'
 export { default as MergeRefs } from './MergeRefs'
-export * from './Popper'
 export { default as Responsive } from './Responsive'