+++ src/core/index.js
@@ -1,3 +1,4 @@
 export { default as getDisplayName } from './getDisplayName'
 export { default as getScreenSize } from './getScreenSize'
+export { default as minWidthModifier } from './minWidthModifier'
 export { default as watchResize } from './watchResize'